+++ src/components/sortable/sortableUtils.ts
@@ -3,18 +3,18 @@ import type { DocId, IntegerId } from '@/types/common'
 export const WIDGET_HTML_ID_PREFIX = 'a-sortable-'
 
 export interface SortableItemDataAware {
+  id?: DocId | IntegerId
   position: number
-  id: DocId | IntegerId
 }
 
 export interface SortableItemWithParentDataAware {
-  position: number
   id: DocId | IntegerId
+  position: number
   parent: DocId | IntegerId | null // if null, no parent
 }
 
 export interface SortableItemNewPosition {
-  id: DocId | IntegerId
+  id?: DocId | IntegerId
   position: number
 }